separate server from its launch to improve modularity and to prep for testing
