+++ .gitignore
@@ -22,6 +22,8 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
+.eslintcache
+
 # editor
 .vscode/*
 !.vscode/settings.sample.json
